remove old thread pool
