+++ src/pool.rs
@@ -1,80 +1,65 @@
-mod local;
 #[cfg(feature = "constellation")]
 mod process;
+mod thread;
 pub(crate) mod util;
 
-pub use local::ThreadPool;
+use futures::future::{BoxFuture, TryFutureExt};
+use std::{error::Error, future::Future};
+
 #[cfg(feature = "constellation")]
 pub use process::ProcessPool;
+pub use thread::ThreadPool;
 
-mod process_pool_impls {
-	use futures::future::{BoxFuture, TryFutureExt};
-	use std::{error::Error, future::Future};
+use amadeus_core::pool::{
+	ProcessPool as ProcessPoolTrait, ProcessSend, ThreadPool as ThreadPoolTrait
+};
 
-	use amadeus_core::pool::{ProcessPool as Pool, ProcessSend};
+type Result<T> = std::result::Result<T, Box<dyn Error + Send>>;
 
-	#[cfg(feature = "constellation")]
-	use super::ProcessPool;
-	use super::ThreadPool;
+#[cfg(feature = "constellation")]
+impl ProcessPoolTrait for ProcessPool {
+	type ThreadPool = ThreadPool;
 
-	type Result<T> = std::result::Result<T, Box<dyn Error + Send>>;
-
-	#[cfg(feature = "constellation")]
-	impl Pool for ProcessPool {
-		type ThreadPool = ThreadPool;
-
-		fn processes(&self) -> usize {
-			ProcessPool::processes(self)
-		}
-		fn spawn<F, Fut, T>(&self, work: F) -> BoxFuture<'static, Result<T>>
-		where
-			F: FnOnce(&Self::ThreadPool) -> Fut + ProcessSend,
-			Fut: Future<Output = T> + 'static,
-			T: ProcessSend,
-		{
-			Box::pin(ProcessPool::spawn(self, work).map_err(|e| Box::new(e) as _))
-		}
+	fn processes(&self) -> usize {
+		ProcessPool::processes(self)
 	}
-
-	impl Pool for ThreadPool {
-		type ThreadPool = Self;
-
-		fn processes(&self) -> usize {
-			1
-		}
-		fn spawn<F, Fut, T>(&self, work: F) -> BoxFuture<'static, Result<T>>
-		where
-			F: FnOnce(&Self::ThreadPool) -> Fut + ProcessSend,
-			Fut: Future<Output = T> + 'static,
-			T: ProcessSend,
-		{
-			let self_ = self.clone();
-			Box::pin(ThreadPool::spawn(self, move || work(&self_)).map_err(|e| Box::new(e) as _))
-		}
+	fn spawn<F, Fut, T>(&self, work: F) -> BoxFuture<'static, Result<T>>
+	where
+		F: FnOnce(&Self::ThreadPool) -> Fut + ProcessSend,
+		Fut: Future<Output = T> + 'static,
+		T: ProcessSend,
+	{
+		Box::pin(ProcessPool::spawn(self, work).map_err(|e| Box::new(e) as _))
 	}
 }
 
-mod thread_pool_impls {
-	use futures::future::{BoxFuture, TryFutureExt};
-	use std::{error::Error, future::Future};
+impl ProcessPoolTrait for ThreadPool {
+	type ThreadPool = Self;
 
-	use amadeus_core::pool::ThreadPool as Pool;
+	fn processes(&self) -> usize {
+		1
+	}
+	fn spawn<F, Fut, T>(&self, work: F) -> BoxFuture<'static, Result<T>>
+	where
+		F: FnOnce(&Self::ThreadPool) -> Fut + ProcessSend,
+		Fut: Future<Output = T> + 'static,
+		T: ProcessSend,
+	{
+		let self_ = self.clone();
+		Box::pin(ThreadPool::spawn(self, move || work(&self_)).map_err(|e| Box::new(e) as _))
+	}
+}
 
-	use super::ThreadPool;
-
-	type Result<T> = std::result::Result<T, Box<dyn Error + Send>>;
-
-	impl Pool for ThreadPool {
-		fn threads(&self) -> usize {
-			ThreadPool::threads(self)
-		}
-		fn spawn<F, Fut, T>(&self, work: F) -> BoxFuture<'static, Result<T>>
-		where
-			F: FnOnce() -> Fut + Send + 'static,
-			Fut: Future<Output = T> + 'static,
-			T: Send + 'static,
-		{
-			Box::pin(ThreadPool::spawn(self, work).map_err(|e| Box::new(e) as _))
-		}
+impl ThreadPoolTrait for ThreadPool {
+	fn threads(&self) -> usize {
+		ThreadPool::threads(self)
+	}
+	fn spawn<F, Fut, T>(&self, work: F) -> BoxFuture<'static, Result<T>>
+	where
+		F: FnOnce() -> Fut + Send + 'static,
+		Fut: Future<Output = T> + 'static,
+		T: Send + 'static,
+	{
+		Box::pin(ThreadPool::spawn(self, work).map_err(|e| Box::new(e) as _))
 	}
 }
